Move hazard xarray tests into separate test module
